rinconadaApp/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
pbkdf2==1.3
pillow==10.3.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.1.0
PythonTurtle==0.3.2
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
twilio==9.1.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.1
Werkzeug==3.0.3
wikipedia==1.4.0
yarl==1.9.4
